fixed a bunch of warnings
git-svn-id: https://bwapi.googlecode.com/svn/trunk@2286 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPILIB/Source/ExplosionType.cpp
@@ -93,7 +93,7 @@ namespace BWAPI
       for(std::set<ExplosionType>::iterator i = explosionTypeSet.begin(); i != explosionTypeSet.end(); i++)
       {
         std::string name=(*i).getName();
-        for(int j=0;j<name.length();j++)
+        for(int j=0;j<(int)name.length();j++)
         {
           if (name[j]==' ') name[j]='_';
           if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
@@ -150,7 +150,7 @@ namespace BWAPI
 
   ExplosionType ExplosionTypes::getExplosionType(std::string& name)
   {
-    for(int j=0;j<name.length();j++)
+    for(int j=0;j<(int)name.length();j++)
     {
       if (name[j]==' ') name[j]='_';
       if (name[j]>='a' && name[j]<='z') name[j]+='A'-'a';
